refactor(github): split BatchResults into reusable components
- Extract RepoCard, RepoActions, RepoStats, and StatItem into separate components
- Create useRepoActions hook for copying URL and opening in GitHub
- Update BatchResults to use new components
- Improve code organization and reusability
